+++ paypal/standard/ipn/models.py
@@ -4,11 +4,7 @@ from __future__ import unicode_literals
 
 import requests
 
-from paypal.standard.ipn.signals import (
-    invalid_ipn_received, payment_was_flagged, payment_was_refunded, payment_was_reversed, payment_was_successful,
-    recurring_cancel, recurring_create, recurring_failed, recurring_payment, recurring_skipped, subscription_cancel,
-    subscription_eot, subscription_modify, subscription_signup, valid_ipn_received
-)
+from paypal.standard.ipn.signals import invalid_ipn_received, valid_ipn_received
 from paypal.standard.models import PayPalStandardBase
 from paypal.utils import warn_untested
 
@@ -33,44 +29,9 @@ class PayPalIPN(PayPalStandardBase):
         """Shout for the world to hear whether a txn was successful."""
         if self.flag:
             invalid_ipn_received.send(sender=self)
-            payment_was_flagged.send(sender=self)
             return
         else:
             valid_ipn_received.send(sender=self)
-
-        # Transaction signals:
-        if self.is_transaction():
-            if self.is_refund():
-                payment_was_refunded.send(sender=self)
-            elif self.is_reversed():
-                payment_was_reversed.send(sender=self)
-            else:
-                payment_was_successful.send(sender=self)
-        # Recurring payment signals:
-        # XXX: Should these be merged with subscriptions?
-        elif self.is_recurring():
-            if self.is_recurring_create():
-                recurring_create.send(sender=self)
-            elif self.is_recurring_payment():
-                warn_untested()
-                recurring_payment.send(sender=self)
-            elif self.is_recurring_cancel():
-                recurring_cancel.send(sender=self)
-            elif self.is_recurring_skipped():
-                recurring_skipped.send(sender=self)
-            elif self.is_recurring_failed():
-                recurring_failed.send(sender=self)
-        # Subscription signals:
-        else:
-            warn_untested()
-            if self.is_subscription_cancellation():
-                subscription_cancel.send(sender=self)
-            elif self.is_subscription_signup():
-                subscription_signup.send(sender=self)
-            elif self.is_subscription_end_of_term():
-                subscription_eot.send(sender=self)
-            elif self.is_subscription_modified():
-                subscription_modify.send(sender=self)
 
     def __repr__(self):
         return '<PayPalIPN id:{0}>'.format(self.id)