+++ paypal/standard/models.py
@@ -364,16 +364,6 @@ class PayPalStandardBase(Model):
                     self.set_flag("Invalid payment_status. (%s)" % self.payment_status)
                 if duplicate_txn_id(self):
                     self.set_flag("Duplicate txn_id. (%s)" % self.txn_id)
-                if hasattr(settings, 'PAYPAL_RECEIVER_EMAIL'):
-                    warn("Use of PAYPAL_RECEIVER_EMAIL in settings has been Deprecated.\n"
-                         "Check of valid email must be done when receiving the\n"
-                         "valid_ipn_received signal",
-                         DeprecationWarning)
-                    if self.receiver_email != settings.PAYPAL_RECEIVER_EMAIL:
-                        self.set_flag("Invalid receiver_email. (%s)" % self.receiver_email)
-            else:
-                # @@@ Run a different series of checks on recurring payments.
-                pass
 
         self.save()